refactor: minor, move function to appropriate file
